Move DeviceCategory display condition to it's own sub-namespace
