+++ src/sprout/sprout_load.php
@@ -55,7 +55,7 @@ Register::cronJob('daily', 'Sprout\\Controllers\\Admin\\FileAdminController', 'c
 Register::cronJob('daily', 'Sprout\\Controllers\\ContentSubscribeController', 'cronSendSubscriptions');
 Register::cronJob('daily', 'Sprout\\Controllers\\Admin\\ActionLogAdminController', 'cronCleanup');
 
-Register::displayCondition('Sprout\\Helpers\\DisplayConditions\\DeviceCategory', 'Platform', 'Device category');
+Register::displayCondition('Sprout\\Helpers\\DisplayConditions\\Platform\\DeviceCategory', 'Platform', 'Device category');
 
 Register::widgetTile(
     'embedded',
